Fixing the vicinity discovery using the ssn improperly.
This commit adds a new event signaling the need to resynchonise with the
given node.
This new event is consumed by the VicinityDiscovery and the
PrecomputePathAndPathIds use-cases.
The VicinityDiscovery use-case manages a queue of nodes to resynchonise
with.
The PrecomputePathAndPathIds use-case is using the event to know which
next valid ssn-range to accept.