Fix pnet dependency missing in default feature.
This is necessary because creating interfaces is now sanity checked on
actual physical existence. This was probably done to enforce valid interfaces
for the forwarding domain.

Rather then enforcing it globally there should be a forwarding specific
interface struct which may be sanity checked.
This also makes it much simpler mocking interfaces for unit tests in the
routing domain.