Start of an _unstable_ branch for integrating new features of KIRA.
This branch will combine merged feature branches.
Features are collected, integrated, tested and fixed on this branch
before a new release is ready to be merged into the main branch.

For feature freezes of releases a dedicated feature branch may be
created where only bug fixes are allowed. The final merge commit onto the
main branch can then be tagged with a version number for a new release.